requirements.txt
openai
azure.cognitiveservices.speech
whisper
bark
scipy
ipython
PIL
soundfile
sounddevice 
torch
ipython
